reinit with cli-3 setup and jest, use rollup for build
