--- .babelrc
@@ -1,21 +1,0 @@
-{
-  "presets": [
-    ["env", {
-      "modules": false,
-      "targets": {
-        "browsers": ["> 1%", "last 2 versions", "not ie <= 8"]
-      }
-    }],
-    "stage-2"
-  ],
-  "plugins": ["transform-vue-jsx", "transform-runtime", "transform-flow-comments"],
-  "env": {
-    "test": {
-      "presets": ["env", "stage-2"],
-      "plugins": ["transform-vue-jsx", ["istanbul", {
-        "include": ["src/**"],
-        "exclude": ["test/**", "src/main.js", "src/flow.types.js"]
-      }]]
-    }
-  }
-}
