--- .flowconfig
@@ -1,12 +1,0 @@
-[ignore]
-.*/node_modules/**/.*
-.*/docs/.*
-.*/test/.*
-.*/config/.*
-
-
-[include]
-
-[libs]
-
-[options]
