+++ .gitignore
@@ -1,5 +1,22 @@
+.DS_Store
 node_modules
-coverage
+
+
+# local env files
+.env.local
+.env.*.local
+
+# Log files
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+
+# Editor directories and files
 .idea
-*.iml
-*.log
+.vscode
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
