--- .npmignore
@@ -1,10 +1,0 @@
-.*
-*.log
-*.swp
-*.yml
-coverage
-docs/_book
-config
-dist/*.map
-lib
-test
