--- config/dev.env.js
@@ -1,5 +1,0 @@
-'use strict'
-
-module.exports = {
-  NODE_ENV: '"development"'
-}
