--- index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <script src="node_modules/es6-promise-polyfill/promise.min.js"></script>
-    <script src="node_modules/es6-shim/es6-shim.min.js"></script>
-    <link rel="shortcut icon" href="https://avatars1.githubusercontent.com/u/1688158?v=4&s=280" />
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <title>molgenis-i18n-js</title>
-  </head>
-  <body>
-    <div id="i18n-demo"></div>
-    <!-- built files will be auto injected -->
-  </body>
-</html>
