+++ package.json
@@ -1,146 +1,48 @@
 {
-  "name": "@molgenis/molgenis-i18n-js",
+  "name": "molgenis-js-i18n",
   "description": "A Vue.js plugin for the internationalization of MOLGENIS Vue applications",
   "version": "1.0.0",
-  "homepage": "https://github.com/molgenis/molgenis-js-i18n#readme",
   "author": {
     "name": "Genomics Coordination Center",
     "email": "gcc-apps-team@googlegroups.com"
-  },
-  "release": {
-    "plugins": [
-      "@semantic-release/commit-analyzer",
-      "@semantic-release/release-notes-generator",
-      [
-        "@semantic-release/changelog",
-        {
-          "changelogFile": "CHANGELOG.md"
-        }
-      ],
-      "@semantic-release/npm",
-      "@semantic-release/git",
-      "@semantic-release/github"
-    ]
   },
   "publishConfig": {
     "access": "public",
     "registry": "http://registry.npmjs.org"
   },
   "main": "dist/molgenis-i18n-js.esm.js",
-  "keywords": [
-    "plugin",
-    "vue",
-    "vuejs",
-    "molgenis",
-    "i18n"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/molgenis/molgenis-js-i18n.git"
-  },
-  "bugs": {
-    "url": "https://github.com/molgenis/molgenis-js-i18n/issues"
-  },
   "license": "LGPL-3.0",
   "scripts": {
-    "dev": "webpack-dev-server --inline --progress --config build/webpack.dev.conf.js",
-    "start": "npm run dev",
-    "unit": "cross-env BABEL_ENV=test karma start test/unit/karma.conf.js --single-run",
-    "test": "npm run unit",
-    "lint": "eslint --ext .js,.vue src test/unit/specs",
-    "build": "node build/build.js",
+    "serve": "vue-cli-service serve",
+    "build": "rollup -c --environment BUILD:production",
+    "unit": "vue-cli-service test:unit",
+    "lint": "vue-cli-service lint",
     "preversion": "yarn install; yarn unit; yarn build"
   },
   "dependencies": {
-    "dayjs": "^1.9.6",
-    "i18next": "^7.1.3",
-    "i18next-xhr-backend": "^1.4.1"
+    "core-js": "^3.6.5",
+    "dayjs": "^1.10.4",
+    "i18next": "^19.8.7",
+    "i18next-http-backend": "^1.1.0",
+    "vue": "^2.6.11"
   },
   "devDependencies": {
-    "@semantic-release/changelog": "^3.0.2",
-    "@semantic-release/git": "^7.0.8",
-    "@vue/test-utils": "^1.0.0-beta.13",
-    "autoprefixer": "^7.1.2",
-    "babel-core": "^6.22.1",
-    "babel-eslint": "^7.1.1",
-    "babel-helper-vue-jsx-merge-props": "^2.0.3",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-istanbul": "^4.1.1",
-    "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-plugin-transform-vue-jsx": "^3.5.0",
-    "babel-preset-env": "^1.3.2",
-    "babel-preset-flow": "^6.23.0",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.22.0",
-    "chai": "^4.1.2",
-    "chalk": "^2.0.1",
-    "chromedriver": "^2.27.2",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-env": "^5.0.1",
-    "cross-spawn": "^5.0.1",
-    "css-loader": "^0.28.0",
-    "es6-promise-polyfill": "^1.2.0",
-    "es6-shim": "^0.35.3",
-    "eslint": "^3.19.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-flowtype": "^2.46.1",
-    "eslint-plugin-html": "^3.0.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.2.0",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^1.1.4",
-    "flow": "^0.2.3",
-    "flow-babel-webpack-plugin": "^1.1.1",
-    "friendly-errors-webpack-plugin": "^1.6.1",
-    "html-webpack-plugin": "^2.30.1",
-    "inject-loader": "^3.0.0",
-    "karma": "^1.4.1",
-    "karma-babel-polyfill": "^0.0.5",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.3.0",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "karma-promise": "^0.1.0",
-    "karma-safari-launcher": "^1.0.0",
-    "karma-sinon-chai": "^1.3.1",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.31",
-    "karma-webpack": "^2.0.2",
-    "mocha": "^3.2.0",
-    "node-notifier": "^5.1.2",
-    "optimize-css-assets-webpack-plugin": "^3.2.0",
-    "ora": "^1.2.0",
-    "phantomjs-prebuilt": "^2.1.14",
-    "portfinder": "^1.0.13",
-    "postcss-import": "^11.0.0",
-    "postcss-loader": "^2.0.8",
-    "raw-loader": "^0.5.1",
-    "rimraf": "^2.6.0",
-    "rollup": "^0.57.1",
-    "rollup-plugin-buble": "^0.19.2",
-    "rollup-plugin-flow-no-whitespace": "^1.0.0",
-    "rollup-plugin-replace": "^2.0.0",
-    "selenium-server": "^3.0.1",
-    "semantic-release": "^17.2.3",
-    "semver": "^5.3.0",
-    "shelljs": "^0.7.6",
-    "sinon": "^4.0.0",
-    "sinon-chai": "^2.8.0",
-    "testdouble": "^3.3.0",
-    "uglifyjs-webpack-plugin": "^1.1.1",
-    "url-loader": "^0.5.8",
-    "vue": "^2.5.13",
-    "vue-loader": "^13.6.2",
-    "vue-style-loader": "^3.0.1",
-    "vue-template-compiler": "^2.5.13",
-    "webpack": "^3.6.0",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-server": "^2.9.1",
-    "webpack-merge": "^4.1.0"
+    "@rollup/plugin-typescript": "^8.1.1",
+    "@types/jest": "^24.0.19",
+    "@typescript-eslint/eslint-plugin": "^2.33.0",
+    "@typescript-eslint/parser": "^2.33.0",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-typescript": "~4.5.0",
+    "@vue/cli-plugin-unit-jest": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/eslint-config-typescript": "^5.0.2",
+    "@vue/test-utils": "^1.0.3",
+    "eslint": "^6.7.2",
+    "eslint-plugin-vue": "^6.2.2",
+    "rollup": "^2.38.5",
+    "rollup-plugin-typescript": "^1.0.1",
+    "typescript": "~3.9.3",
+    "vue-template-compiler": "^2.6.11"
   }
 }
